Add verification endpoint sending simple, hardcoded mail.
